fully implemented parser
parser.py parses the tokenized mini code generated from lex.py. It then implements the logic and grammer of the mini code. This allows for proper handling of operator precedence. While parsing, emit.py functions are used to output the compiled mini code into C code which is generated and written to out.c in the same file directory. The out.c file can be compliled using gcc and be run as expected